requirements.txt
# scilit-agent — requirements.txt
# Install with: pip install -r requirements.txt

# ── Core ──────────────────────────────────────────────────────
python-dotenv>=1.0.0
pydantic>=2.0.0
tqdm>=4.66.0
numpy>=1.26.0

# ── LLM providers ─────────────────────────────────────────────
langchain>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langchain-huggingface>=0.0.3
openai>=1.30.0
anthropic>=0.25.0

# ── LangGraph ──────────────────────────────────────────────────
langgraph>=0.1.0

# ── Sources ───────────────────────────────────────────────────
arxiv>=2.1.0
biopython>=1.83
requests>=2.31.0
semanticscholar>=0.8.0
watchdog>=4.0.0

# ── Parsers ───────────────────────────────────────────────────
PyMuPDF>=1.24.0
marker-pdf>=0.2.6
lxml>=5.0.0
thefuzz>=0.22.0
python-Levenshtein>=0.25.0   # speeds up thefuzz

# ── Retrieval ─────────────────────────────────────────────────
chromadb>=0.5.0
sentence-transformers>=3.0.0
rank-bm25>=0.2.2

# ── Scheduler ─────────────────────────────────────────────────
APScheduler>=3.10.0

# ── Interfaces ────────────────────────────────────────────────
gradio>=4.0.0
PyQt6>=6.7.0

# ── Dev / test ────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
